some minor reflow of code
